Added weapon types. Reworked gun class and renamed it to turret Moved the generic parts of the update function to the "base" class Added the hull class, integration is still WIP because I need to get shields done.
TODO: Start adding dedicated params for required values.
(so kwargs are used for optional config only)
Make it so the gameworld can be accessed by everything contained by it.
(so I can just tell something to shoot and not have to worry
about adding the bullets to the gameworld externally to the class)

Documentation will start once I'm satisfied with the game structure.